requirements_copied_histories.txt
requests==2.32.4
PyGithub==2.6.1
